prepare for better imports in tests
