+++ .editorconfig
@@ -5,6 +5,10 @@ insert_final_newline = true
 trim_trailing_whitespace = true
 end_of_line = lf
 
+[*.ts]
+indent_style = tab
+indent_size = 4
+
 [*.{md,json,jsonc}]
 indent_style = space
 indent_size = 2
